+++ crates/ruma-common/CHANGELOG.md
@@ -29,6 +29,7 @@ Improvements:
 * Add support for knocking in `events::room::member::MembershipChange`
 * Add `MatrixVersion::V1_3`
 * Deprecate the `sender_key` and `device_id` fields for encrypted events (MSC3700)
+* Move the `relations` field of `events::unsigned` types out of `unstable-msc2675`
 
 # 0.9.2